Generate PDF from PNG using Imagemagick
Generate the board PDFs from the rasterized PNG files using "convert"
(from Imagemagick) rather than directly from the SVG using Inkscape.
Looking at the PDFs with Okular seemed to show some slight differences
in the positions of outlines and fills, and I'm not sure whether that is
a rendering issue in Okular or a PDF export issue in Inkscape.
Converting the PNG files to PDF sidesteps the problem nicely though.

Signed-off-by: James Hogan <james@albanarts.com>