Move tickets up page slightly
Move tickets up the page by 6mm to avoid larger print margins at the
bottom. On HP Deskjet 450 the upper margin is 2mm and the lower 13.5,
which doesn't quite fit 5 landscape tickets, but by adjusting the
position we can reduce the overrun to 0.5mm at top and bottom.

Arguably we should switch to landscape and try and fit 9 per page
instead of 10.

Signed-off-by: James Hogan <james@albanarts.com>