README: Clarify sources and license as CC BY-SA 3.0
Clarify the licenses of the sources in the README, which are all either
public domain or CC BY-SA 3.0 (possibly among others), grouping them by
license a bit clearer and summarising the attribution.

This means that the whole project can be safely published under the CC
BY-SA 3.0 license, so make that clear in the README and update the
copyright year to 2020 since it was never released in 2019.

Signed-off-by: James Hogan <james@albanarts.com>