Makefile: Tile more reliably
The tiling behaviour of pdfposter was a bit unreliable depending on the
exact page size specified, since we used a fixed scaling factor of 1.0
and a slightly arbitrary page size to try and get it to behave sensibly
without any spare space around the pages. Due to the varying DPI however
the sizes of the images are slightly different which results in the
75dpi one tiling A4 portrait.

For a more reliable result, update the Makefile to retrieve the desired
dimentions from the board SVG, perform some arithmetic (using bc) to
calculate the number of tiles in each direction and the size of each
tile, and then pass that through to pdfposter along with the total
poster size (rather than the scale factor). The tile sizes are also
substituted into the tex files so the resulting size on the final PDF
pages is correct.

Signed-off-by: James Hogan <james@albanarts.com>